Version 0.0.1
Decorators are used to give users flexibility to define their own wrappers for output. All decorators my extend SafeEchoDecorator.

3 built-in decorators are included.
NoWrapDecorator - as the default decorator.
MouseHoverDecroator - as an easy to understand, generic use case.
ApiDecryptionDecorator - as a complex use case.

SafeEchoEntityWrapper is a class wrapper that overrides the wrapped class's output and attempts to safeecho (see SafeWrapper for sample implementation). It is highly recommended that a decorator be configured in conjunction with SafeEchoEntityWrapper. Its current implementation may not be viable, since it was part of a scrapped draft. It was left here to provoke thought.

The ApiDecryptionStripper is included to show basic concepts in respect to the ApiDecryptionDecorator. It may never be used. But it is a reference piece for the flip side of ApiDecryptionDecorator.

Consult the README.md for usage specifications.